feat：1.agent流式对话 2.前端行程规划 3.前端接口鉴权 4.后端对话历史

+++ backend/src/main/java/com/asta/backend/entity/query/PageQuery.java
@@ -11,8 +11,8 @@ import lombok.ToString;
 public class PageQuery {
 
     @Min(value = 1, message = "页码最小值为1")
-    private long pageIndex;
+    private int pageIndex;
 
     @Min(value = 1, message = "条数最小值为1")
-    private long pageSize;
+    private int pageSize;
 }
